Add modifications to graphql_query_derive
