+++ examples/example_module/src/lib.rs
@@ -12,6 +12,6 @@ use custom_scalars::*;
 #[graphql(
     schema_path = "../github/src/schema.graphql",
     query_path = "../github/src/query_1.graphql",
-    response_derives = "Debug",
+    response_derives = "Debug"
 )]
 pub struct ExampleModule;
