+++ graphql_client_codegen/src/lib.rs
@@ -4,6 +4,9 @@
 extern crate failure;
 extern crate graphql_parser;
 extern crate heck;
+extern crate itertools;
+#[macro_use]
+extern crate lazy_static;
 extern crate proc_macro;
 extern crate proc_macro2;
 extern crate serde;
@@ -43,6 +46,15 @@ use heck::SnakeCase;
 mod tests;
 use proc_macro2::{Ident, Span};
 
+type CacheMap<T> =
+    ::std::sync::Mutex<::std::collections::hash_map::HashMap<::std::path::PathBuf, T>>;
+
+lazy_static! {
+    static ref SCHEMA_CACHE: CacheMap<schema::Schema> = CacheMap::default();
+    static ref QUERY_CACHE: CacheMap<(String, graphql_parser::query::Document)> =
+        CacheMap::default();
+}
+
 pub struct GraphQLClientDeriveOptions<'a> {
     pub input: &'a syn::DeriveInput,
 }
@@ -66,27 +78,48 @@ pub fn generate_module_token_stream(
         .unwrap_or(deprecation::DeprecationStrategy::Warn);
 
     // We need to qualify the query with the path to the crate it is part of
-    let query_string = read_file(&query_path)?;
-    let query = graphql_parser::parse_query(&query_string)?;
-
-    // We need to qualify the schema with the path to the crate it is part of
-    let schema_string = read_file(&schema_path)?;
-
-    let extension = schema_path
-        .extension()
-        .and_then(|e| e.to_str())
-        .unwrap_or("INVALID");
-
-    let schema = match extension {
-        "graphql" | "gql" => {
-            let s = graphql_parser::schema::parse_schema(&schema_string)?;
-            schema::Schema::from(s)
+    let (query_string, query) = {
+        let mut lock = QUERY_CACHE.lock().expect("query cache is poisoned");
+        match lock.entry(query_path) {
+            ::std::collections::hash_map::Entry::Occupied(o) => o.get().clone(),
+            ::std::collections::hash_map::Entry::Vacant(v) => {
+                let query_string = read_file(v.key())?;
+                let query = graphql_parser::parse_query(&query_string)?;
+                v.insert((query_string, query)).clone()
+            }
         }
-        "json" => {
-            let parsed: FullResponse<introspection_response::IntrospectionResponse> = ::serde_json::from_str(&schema_string)?;
-            schema::Schema::from(parsed.data)
+    };
+
+    // Check the schema cache.
+    let schema = {
+        let mut lock = SCHEMA_CACHE.lock().expect("schema cache is poisoned");
+        match lock.entry(schema_path) {
+            ::std::collections::hash_map::Entry::Occupied(o) => o.get().clone(),
+            ::std::collections::hash_map::Entry::Vacant(v) => {
+                let schema_string = read_file(v.key())?;
+                let schema = {
+                    let extension = v
+                        .key()
+                        .extension()
+                        .and_then(|e| e.to_str())
+                        .unwrap_or("INVALID");
+
+                    match extension {
+                        "graphql" | "gql" => {
+                            let s = graphql_parser::schema::parse_schema(&schema_string)?;
+                            schema::Schema::from(s)
+                        }
+                        "json" => {
+                            let parsed: FullResponse<introspection_response::IntrospectionResponse> = ::serde_json::from_str(&schema_string)?;
+                            schema::Schema::from(parsed.data)
+                        }
+                        extension => panic!("Unsupported extension for the GraphQL schema: {} (only .json and .graphql are supported)", extension)
+                    }
+                };
+
+                v.insert(schema).clone()
+            }
         }
-        extension => panic!("Unsupported extension for the GraphQL schema: {} (only .json and .graphql are supported)", extension)
     };
 
     let module_name = Ident::new(&input.ident.to_string().to_snake_case(), Span::call_site());
@@ -129,20 +162,18 @@ pub fn generate_module_token_stream(
     Ok(result)
 }
 
-fn read_file(
-    path: impl AsRef<::std::path::Path> + ::std::fmt::Debug,
-) -> Result<String, failure::Error> {
+fn read_file(path: &::std::path::Path) -> Result<String, failure::Error> {
     use std::io::prelude::*;
 
     let mut out = String::new();
-    let mut file = ::std::fs::File::open(&path).map_err(|io_err| {
+    let mut file = ::std::fs::File::open(path).map_err(|io_err| {
         let err: failure::Error = io_err.into();
         err.context(format!(
             r#"
-            Could not find file with path: {:?}
+            Could not find file with path: {}
             Hint: file paths in the GraphQLQuery attribute are relative to the project root (location of the Cargo.toml). Example: query_path = "src/my_query.graphql".
             "#,
-            path
+            path.display()
         ))
     })?;
     file.read_to_string(&mut out)?;
