+++ graphql_query_derive/Cargo.toml
@@ -10,14 +10,6 @@ repository = "https://github.com/tomhoule/graphql-client"
 proc-macro = true
 
 [dependencies]
-failure = "0.1"
-itertools = "0.7"
-lazy_static = "1.0"
-quote = "^0.6"
 syn = "0.14"
 proc-macro2 = { version = "0.4", features = [] }
-serde = "1.0"
-serde_derive = "1.0"
-serde_json = "1.0"
-heck = "0.3"
-graphql-parser = "0.2"
+graphql_client_codegen = { path = "../graphql_client_codegen/" }
